pyproject.toml
[tool.poetry]
name = "pay2u"
version = "0.1.0"
description = ""
authors = ["Denis Tretyakov <den.tretyakoff@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.15.0"
django-filter = "^24.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
taskipy = "^1.12.2"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
makemigrations = "python backend/manage.py makemigrations"
migrate = "python backend/manage.py migrate"
startdb = "docker compose -f infra/postgres-local.yaml up -d"
stopdb = "docker compose -f infra/postgres-local.yaml down"
onlydb = "make -f infra/db_only start"